[feat] admin model migration added

+++ app/Http/Controllers/Backend/OtpCodeController.php
@@ -9,7 +9,6 @@ use App\Http\Services\Backend\OtpCodeService;
 use App\Models\VerificationCode;
 use App\Traits\Common\ModelAction;
 use App\Traits\Common\ModelProperty;
-use Exception;
 use Illuminate\Http\RedirectResponse;
 use Inertia\Response;
 
@@ -17,23 +16,20 @@ class OtpCodeController extends Controller
 {
     use ModelAction ,ModelProperty;
 
-    protected array $modelProperty ;
-
+    protected array $modelProperty;
 
     /**
      * Constructor to inject services and apply middleware
      */
     public function __construct(
         protected OtpCodeService $service,
-    )
-    {
+    ) {
         $this->modelProperty = $this->getCommonProperty(
-            resourcePagePrefix :'OtpCode',
-			routePrefix: 'backend.otp-codes'
-		);
-        
-        $this->authorizeResource(VerificationCode::class);
+            resourcePagePrefix : 'OtpCode',
+            routePrefix: 'backend.otp-codes'
+        );
 
+        // $this->authorizeResource(VerificationCode::class, 'otp_code');
     }
 
     /**
@@ -41,43 +37,44 @@ class OtpCodeController extends Controller
      */
     public function index(): Response
     {
+        $user = auth()->user();
+        if (! $user->can('view', VerificationCode::class)) {
+            abort(403, 'Unauthorized action.');
+        }
+
         $data = formatResourceResponse(
             $this->service->getAllCode(),
             OtpCodeResource::class
         );
 
         return AppResponse::asSuccess()
-                    ->withComponent($this->modelProperty['pagePrefix'].'Index', [
-                            'title'                => translate('OTP Codes'),
-                            'data'                 => $data,
-                            'modelProperty'        => $this->modelProperty,
-                            'advanceFilterOptions' => $this->service->getAdvanceFilterOptions()
-                    ])->build();
+            ->withComponent($this->modelProperty['pagePrefix'].'Index', [
+                'title' => translate('OTP Codes'),
+                'data' => $data,
+                'modelProperty' => $this->modelProperty,
+                'advanceFilterOptions' => $this->service->getAdvanceFilterOptions(),
+            ])->build();
     }
 
-   
     /**
      * Summary of destroy
-     * @param VerificationCode $otpCode
-     * @return RedirectResponse
      */
     public function destroy(VerificationCode $otpCode): RedirectResponse
     {
-
+        $user = auth()->user();
+        if (! $user->can('delete', $otpCode)) {
+            abort(403, 'Unauthorized action.');
+        }
         try {
-
             $this->service->delete($otpCode->id);
-            return AppResponse::asSuccess()
-                                ->withMessage('OTP Code deleted successfully.')
-                                ->build();
 
+            return AppResponse::asSuccess()
+                ->withMessage('OTP Code deleted successfully.')
+                ->build();
         } catch (\Exception $e) {
             return AppResponse::asError()
-                                ->withMessage($e->getMessage())
-                                ->build();
+                ->withMessage($e->getMessage())
+                ->build();
         }
     }
-
-
-    
 }
